+++ docs/CHANGELOG.md
@@ -20,6 +20,15 @@ We generally fold multiple commits pertaining to the same topic as a single entr
 - Read the `Breaking Changes` section (here in the changelog).
 - Please report any issue!
 
+## VERSION 0.8.x
+
+### New
+* added `set_init_file(...)` #1073
+* added `load_init_file(...)` #1073
+* added `save_init_file(...)` #1073
+* added `no_saved_settings` keyword to window
+* added `no_saved_settings` keyword to table
+
 ## VERSION 0.8.39
 
 ### Fixes